Finalize the algorithm

+++ src/Node.java
@@ -13,8 +13,8 @@ public class Node  {
     public boolean isTraversable;
     public int position_X;
     public int position_Y;
-    public int g_cost;
-    public int h_cost;
+    public double g_cost;
+    public double h_cost;
     public Node parent;
 
     public Node(boolean traversable, int position_X, int position_Y) {
@@ -23,7 +23,7 @@ public class Node  {
         this.position_Y = position_Y;
     }
 
-    public int f_cost() {
+    public double f_cost() {
         return g_cost + h_cost;
     }
